+++ resources/examples/persistent.js
@@ -2,9 +2,10 @@
 // @macro:persistent
 
 var invocationCount;
-invocationCount ||= 0;
 
 async function main(invocation) {
+  invocationCount ||= 0;
+
   await vscode.window.showInformationMessage(
     `Hello from call ${invocation}. Run this macro a couple of times before dismissing this message.`);
   await vscode.window.showInformationMessage(